refac: rename project
