+++ pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "argstash"
+name = "fdstash"
 version = "0.1.0"
 description = "Add your description here"
 readme = "README.md"
