+++ src/fdstash/backend_base.py
