+++ src/fdstash/backend_inline.py
