+++ src/fdstash/backend_mem.py
