+++ src/fdstash/exceptions.py
