+++ src/fdstash/stash.py
